Creation of a master stats dictionary
Shifted a load of information about each of the stats into a single dictionary in the team section. Technically better and more resilient, but has no meaningful effect on how the code works, sadly.
Also reduced text size on live game tab

+++ frisbee_match.py
@@ -280,7 +280,7 @@ class FrisbeeGame():
         message_string = ""
         for player in self.point_lineups[reference_point_number]:
             message_string = message_string + player + " // "
-        wind_direction_indicator = (self.wind_direction + 3) / 2
+        wind_direction_indicator = int((self.wind_direction + 3) / 2)
         message_string = message_string + "Wind direction " + str(wind_direction_indicator)
         messagebox.showinfo(title=messagebox_title, message=message_string)
